Refactored code to allow for load to be stopped
