+++ ElasticLoadGenerator/ElasticLoadGenerator.exe.config
@@ -3,8 +3,8 @@
   
   <appSettings>
     <!-- Default Screen Values -->
-    <add key="PrimaryDatabase" value="Customer1" />
-    <add key="SecondaryDatabase" value="Customer2" />
+    <add key="PrimaryDatabase" value="WingTipTicketsTenantPri" />
+    <add key="SecondaryDatabase" value="WingTipTicketsTenantSec" />
     <add key="Username" value="developer" />
     <add key="Password" value="P@ssword1" />
     <add key="BatchSize" value="50" />
@@ -17,8 +17,8 @@
     <add key="CustomerName" value="Generator" />
 
     <!-- Timing in Seconds -->
-    <add key="Runtime" value="10800" /> <!-- 300 -->
-    <add key="Sleeptime" value="300" /> <!-- 10 -->
+    <add key="Runtime" value="300" /> <!-- 1080 -->
+    <add key="Sleeptime" value="10" /> <!-- 300 -->
     
     <!-- Database settings -->
     <add key="TransientFaultHandlingRetryCount" value="30" />
