+++ ElasticLoadGenerator/ElasticLoadGenerator.vshost.exe.config
@@ -17,8 +17,8 @@
     <add key="CustomerName" value="Generator" />
 
     <!-- Timing in Seconds -->
-    <add key="Runtime" value="10800" /> <!-- 300 -->
-    <add key="Sleeptime" value="300" /> <!-- 10 -->
+    <add key="Runtime" value="300" /> <!-- 1080 -->
+    <add key="Sleeptime" value="10" /> <!-- 300 -->
     
     <!-- Database settings -->
     <add key="TransientFaultHandlingRetryCount" value="30" />
